(Feat) store extra question
