+++ src/features/resume/download/api/fetchQuestion.ts
@@ -1,16 +1,6 @@
 import { safePost } from '@/src/shared/api/request';
 import { useAuthStore } from '@/src/entities/auth/model/auth';
-import { create } from 'zustand';
-
-interface AIResponseState {
-  isLoading: boolean;
-  setIsLoading: (loading: boolean) => void;
-}
-
-export const useAIResponseState = create<AIResponseState>((set) => ({
-  isLoading: false,
-  setIsLoading: (loading) => set({ isLoading: loading }),
-}));
+import { useAIResponseState } from '@/src/features/resume/download/model/extraQuestionStore';
 
 export const fetchQuestion = async () => {
   const { setIsLoading } = useAIResponseState.getState();
