+++ src/shared/model/useSSE.ts
@@ -3,7 +3,7 @@
 import { useAuthStore } from '@/src/entities/auth/model/auth';
 import { retryWithRefreshedToken } from '@/src/entities/auth/lib/authManager';
 
-import { eventAIQuestionReady } from './eventAIQuestionReady';
+import { eventExtraQuestion } from '@/src/features/resume/download/model/eventExtraQuestion';
 import { eventResumeUpload } from '@/src/features/resume/upload/model/eventResumeUpload';
 
 import { useRef } from 'react';
@@ -49,7 +49,7 @@ export const useSSE = () => {
         }
       });
 
-      eventAIQuestionReady(eventSource);
+      eventExtraQuestion(eventSource);
       eventResumeUpload(eventSource);
 
       eventSource.onerror = (error) => {
